+++ wp-content/plugins/superb-blocks/plugin.php
@@ -6,14 +6,14 @@ namespace SuperbAddons;
 Plugin Name: Superb Addons: Blocks, Patterns & Theme Designer
 Plugin URI: http://superbthemes.com/
 Description: Superb Addons: Blocks, Patterns & Theme Designer
-Version: 3.5.7
+Version: 3.5.9
 Author: SuperbThemes
 Author URI: http://superbthemes.com/
 License: GPL-3.0+
 License URI: https://www.gnu.org/licenses/gpl-3.0.txt
 Text Domain: superb-blocks
-* Elementor tested up to: 3.27
-* Elementor Pro tested up to: 3.27
+* Elementor tested up to: 3.28
+* Elementor Pro tested up to: 3.28
 */
 
 defined('ABSPATH') || exit;
@@ -23,7 +23,7 @@ if (!defined('WPINC')) {
 }
 // Constants
 if (!defined('SUPERBADDONS_VERSION')) {
-    define('SUPERBADDONS_VERSION', '3.5.7');
+    define('SUPERBADDONS_VERSION', '3.5.9');
 }
 
 if (!defined('SUPERBADDONS_LIBRARY_VERSION')) {