+++ wp-content/plugins/superb-blocks/src/components/slots/class-css-blocks-base-slot.php
@@ -21,8 +21,7 @@ class CssBlocksBaseSlot extends PremiumSlot
 <?php
             },
             array("superbaddons-element-ml1"),
-            AdminLinkSource::CSS_EXPORT,
-            ['url' => 'https://superbthemes.com/features/advanced-custom-css/']
+            AdminLinkSource::CSS_EXPORT
         );
     }
 }