+++ wp-content/plugins/superb-blocks/src/components/slots/class-css-blocks-export-selected-slot.php
@@ -21,8 +21,7 @@ class CssBlocksExportSelectedSlot extends PremiumSlot
 <?php
             },
             array(),
-            AdminLinkSource::CSS_EXPORT,
-            ['url' => 'https://superbthemes.com/features/advanced-custom-css/']
+            AdminLinkSource::CSS_EXPORT
         );
     }
 }