+++ wp-content/plugins/superb-blocks/src/data/utils/class-addons-template-util.php
@@ -64,19 +64,18 @@ class AddonsPageTemplateUtil
 
     public static function GetAddonsPageTemplateContent($page_content)
     {
-        $template = <<<HTML
-        <!-- wp:template-part {"slug":"header","lock":{"move":true,"remove":true}} /-->
-        <!-- wp:group {"tagName":"main","lock":{"move":true,"remove":true},"style":{"spacing":{"margin":{"top":"0rem"},"padding":{"top":"0","bottom":"0","left":"0","right":"0"},"blockGap":"var:preset|spacing|superbspacing-medium"}},"layout":{"type":"default"}} -->
-        <main class="wp-block-group" style="margin-top:0rem;padding-top:0;padding-right:0;padding-bottom:0;padding-left:0">
-HTML;
+        $template = '<!-- wp:template-part {"slug":"header","lock":{"move":true,"remove":true}} /-->' . "\n";
+        $template .= '<!-- wp:group {"tagName":"main","lock":{"move":true,"remove":true},"style":{"spacing":{"margin":{"top":"0rem"},"padding":{"top":"0","bottom":"0","left":"0","right":"0"},"blockGap":"var:preset|spacing|superbspacing-medium"}},"layout":{"type":"default"}} -->' . "\n";
+        $template .= '<main class="wp-block-group" style="margin-top:0rem;padding-top:0;padding-right:0;padding-bottom:0;padding-left:0">' . "\n";
+
         AllowedTemplateHTMLUtil::enable_safe_styles();
         $template .= wp_kses($page_content, "post");
         AllowedTemplateHTMLUtil::disable_safe_styles();
-        $template .= <<<HTML
-        </main>
-        <!-- /wp:group -->
-        <!-- wp:template-part {"slug":"footer","lock":{"move":true,"remove":true}} /-->
-HTML;
+
+        $template .= '</main>' . "\n";
+        $template .= '<!-- /wp:group -->' . "\n";
+        $template .= '<!-- wp:template-part {"slug":"footer","lock":{"move":true,"remove":true}} /-->';
+
         return $template;
     }
 }