+++ wp-content/plugins/superb-blocks/src/data/utils/wizard/class-wizard-template-provider.php
@@ -73,7 +73,6 @@ class WizardTemplateProvider
 
     private function InitializePart($slugOrWizardItem, $type, $regularTitle, $modifiedTitle)
     {
-
         if ($slugOrWizardItem instanceof WizarditemRestorationPoint) {
             $this->AddToAppropriateArray($slugOrWizardItem, $slugOrWizardItem->GetBaseSlug());
             return;