+++ wp-content/plugins/superb-blocks/src/elementor/controllers/class-elementor-controller.php
@@ -56,7 +56,7 @@ class ElementorController
         return array(
             array(
                 "id" => "sections",
-                "premium_url" => AdminLinkUtil::GetLink(AdminLinkSource::LIBRARY_ITEM, ["url" => "https://superbthemes.com/features/patterns/"]),
+                "premium_url" => AdminLinkUtil::GetLink(AdminLinkSource::LIBRARY_ITEM),
                 "title" => esc_html__('Sections', "superb-blocks"),
                 "routes" => array(
                     "list" => LibraryRequestController::ELEMENTOR_LIST_ROUTE,
@@ -79,8 +79,6 @@ class ElementorController
                 "insert_error" => esc_html__('Something went wrong while attempting to insert this element. Please try again or contact support if the problem persists.', "superb-blocks"),
                 "list_error" => esc_html__('Something went wrong while attempting to list elements. Please try again or contact support if the problem persists.', "superb-blocks")
             ),
-            'premium_page_url' => 'https://superbthemes.com/features/page-library/',
-            'premium_pattern_url' => 'https://superbthemes.com/features/patterns/',
             "menu_items" => self::GetElementorLibraryMenuItems(),
             "rest" => array(
                 "base" => \get_rest_url(),