+++ wp-content/plugins/hostinger/changelog.md
@@ -408,3 +408,9 @@ Changelog
 
 3.0.34 (2024-04-10)
 - Add preview and site edit buttons
+
+3.0.35 (2024-04-28)
+- Add edit site button support for block themes
+
+3.0.36 (2024-04-29)
+- Style corrections